// projects/shell/src/app/shared/models/customer/individual-formstate.ts
export interface AdditionalPersonalDetails {
  language?: string;
  dependants?: string;
  qualification?: string;
  institutionName?: string;
  rentalStatus?: string;
  socialStatus?: string;
  spouseName?: string;
  socialEconomicLevel?: string;
  averageMonthlyExpenditure?: string;
  spouses?: {
    spousesFullName?: string;
    spouseIdentificationNumber?: string;
    spouseIdentificationType?: string;
    maidenName?: string;
  }[];
  smrCode?: string;
  fateStatus?: string;
  disability?: string;
  numberOfSpouses?: string;
  negativeClientStatus?: string;
}